+++ src/realm/table.cpp
@@ -2109,6 +2109,7 @@ template ObjKey Table::find_first(ColKey col_key, ObjKey) const;
 template ObjKey Table::find_first(ColKey col_key, util::Optional<bool>) const;
 template ObjKey Table::find_first(ColKey col_key, util::Optional<int64_t>) const;
 template ObjKey Table::find_first(ColKey col_key, BinaryData) const;
+template ObjKey Table::find_first(ColKey col_key, util::Optional<ObjectId>) const;
 
 ObjKey Table::find_first_int(ColKey col_key, int64_t value) const
 {